Move `DynatraceOAuthClient.ts` to the common folder to shared use
